--- src/panes/Info.css
@@ -1,21 +1,0 @@
-.node .sharing {
-  padding: 0 1em 1em;
-}
-
-.node .sharing p {
-  margin: 0;
-}
-
-.node .sharing p:first-of-type {
-  font-size: 35px;
-}
-
-.node .sharing p:last-of-type {
-  font-size: 18px;
-  color: rgba(255,255,255,0.5);
-}
-
-.light .node .sharing p:last-of-type {
-  font-size: 18px;
-  color: rgba(0,0,0,0.5);
-}
